refactor: Reduce TriggerManager file size
extract DumbTrigger Manager to its own file
export default legacyTriggerManager on the same line in order to see
directly what is the default exported component
