+++ packages/cozy-harvest-lib/src/components/KonnectorModal.jsx
@@ -20,9 +20,9 @@ import Infos from 'cozy-ui/transpiled/react/deprecated/Infos'
 
 import AccountSelectBox from './AccountSelectBox/AccountSelectBox'
 import AccountsList from './AccountsList/AccountsList'
+import TriggerManager from './DumbTriggerManager'
 import DataTab from './KonnectorConfiguration/DataTab'
 import KonnectorAccountWrapper from './KonnectorConfiguration/KonnectorAccountWrapper'
-import TriggerManager from './TriggerManager'
 import KonnectorUpdateInfos from './infos/KonnectorUpdateInfos'
 import { fetchAccount } from '../connections/accounts'
 import { fetchTrigger } from '../connections/triggers'
