+++ packages/cozy-harvest-lib/src/components/TriggerManager.spec.jsx
@@ -1,6 +1,6 @@
 /* eslint-env jest */
 import { render, fireEvent, cleanup, screen } from '@testing-library/react'
-import { DumbTriggerManager as TriggerManager } from 'components/TriggerManager'
+import { DumbTriggerManager as TriggerManager } from 'components/DumbTriggerManager'
 import omit from 'lodash/omit'
 import React from 'react'
